requirements.txt
fastapi
uvicorn
azure-storage-file-datalake
python-dotenv
python-multipart
requests
aiofiles